worked on encoding implementation

+++ rand.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
 int main() {
-	printf("%li\n", sizeof(short)*8);
+	printf("%li\n", sizeof(long int)*8);
 	return 0;
 }
